feat: add negative values and refactor
