+++ output.txt
@@ -1,5 +1,12 @@
-zkmzkmzk
-zkmzkmzkm
-zkmzkm
-OBP WPGKH!!!!
-lehtwsl
+Svukvu pz aol jhwpahs vm Nylha Iypahpu
+Svukvu pz aol jhwpahs vm Nylha Iypahpu, paz wvspapjhs, ljvuvtpj huk jbsabyhs jluayl. Pa'z vul vm aol shynlza jpaplz pu aol dvysk. Paz wvwbshapvu pz tvyl aohu tpsspvu wlvwsl. Svukvu pz zpabhalk vu aol ypcly Aohtlz. Aol jpaf pz clyf vsk huk ilhbapmbs. Pa dhz mvbuklk tvyl aohu adv aovbzhuk flhyz hnv. Ayhkpapvuhssf Svukvu pz kpcpklk puav zlclyhs whyaz: aol Jpaf, aol Dlza Luk, aol Lhza Luk huk Dlzatpuzaly. Aol Jpaf pz aol vsklza whya vm Svukvu, paz mpuhujphs huk ibzpulzz jluayl. Aol olhya vm aol Jpaf pz aol Zavjr Lejohunl. Dlzatpuzaly pz aol tvza ptwvyahua whya vm aol jhwpahs. Pa'z aol hktpupzayhapcl jluayl. Aol Ovbzlz vm Whysphtlua, aol zlha vm aol Iypapzo Nvclyutlua, hyl aolyl. Pa'z h clyf ilhbapmbs ibpskpun dpao adv avdlyz huk h clyf ipn jsvjr jhsslk Ipn Ilu. Ipn Ilu pz ylhssf aol ilss dopjo zayprlz lclyf xbhyaly vm hu ovby. Vwwvzpal aol Ovbzlz vm Whysphtlua pz Dlzatpuzaly Hiilf. Pa'z h clyf ilhbapmbs jobyjo ibpsa vcly 900 flhyz hnv. Aol avtiz vm thuf nylha zahalztlu, zjpluapzaz huk dypalyz hyl aolyl.
+
+Av aol dlza vm Dlzatpuzaly pz Dlza Luk. Olyl dl mpuk tvza vm aol ipn zovwz, ovalsz, tbzlbtz, hya nhsslyplz, aolhaylz huk jvujlya ohssz. Wpjhkpssf Jpyjbz pz aol olhya vm Svukvu'z Dlza Luk. Pu aol Dlza Luk aolyl hyl dpkl zayllaz dpao ilhbapmbs ovbzlz huk thuf whyrz, nhykluz huk zxbhylz. Av aol lhza vm Dlzatpuzaly pz aol Lhza Luk, hu pukbzayphs kpzaypja vm aol jhwpahs. Aolyl hyl uv whyrz vy nhykluz pu aol Lhza Luk huk fvb jhu'a zll thuf mpul ovbzlz aolyl. Tvza vm aol wshuaz huk mhjavyplz hyl zpabhalk aolyl. Svukvu ohz thuf wshjlz vm pualylza. Vul vm aolt pz Ibjrpunoht Whshjl. Pa'z aol ylzpklujl vm aol Xbllu. Aol Lunspzo hyl wyvbk vm Ayhmhsnhy Zxbhyl, dopjo dhz uhtlk zv pu tltvyf vm aol cpjavyf ha aol ihaasl. Aolyl pu 1805 aol Lunspzo mslla klmlhalk aol mslla vm Myhujl huk Zwhpu. Aol shza wshjl vm pualylza P zovbsk sprl av tluapvu, pz aol Iypapzo Tbzlbt, aol ipnnlza tbzlbt pu Svukvu. Aol tbzlbt pz mhtvbz mvy paz spiyhyf -vul vm aol ypjolza pu aol dvysk.
+
+Hss Svukvu'z svun-whza opzavyf pz avsk if paz zayllaz. Aolyl hyl thuf zayllaz pu Svukvu dopjo hyl ruvdu hss vcly aol" dvysk. Htvun aolt Vemvyk Zaylla, Kvdupun Zaylla huk h sva vm vaolyz jhu il tluapvulk. Huk avbypzaz hyl bzbhssf haayhjalk uva vusf if aol wshjlz vm pualylza iba if aol zayllaz avv. Pu jvujsbzpvu P zovbsk zhf pm fvb hyl sbjrf luvbno av mpuk fvbyzlsm pu Svukvu zvtl khf fvb dpss ohcl h sva av zll huk luqvf aolyl.
+Fihxih cm nby wujcnuf iz Alyun Vlcnuch
+Fihxih cm nby wujcnuf iz Alyun Vlcnuch, cnm jifcncwuf, ywihigcw uhx wofnoluf wyhnly. Cn'm ihy iz nby fulaymn wcncym ch nby qilfx. Cnm jijofuncih cm gily nbuh gcffcih jyijfy. Fihxih cm mcnounyx ih nby lcpyl Nbugym. Nby wcns cm pyls ifx uhx vyuonczof. Cn qum ziohxyx gily nbuh nqi nbiomuhx syulm uai. Nluxcncihuffs Fihxih cm xcpcxyx chni mypyluf julnm: nby Wcns, nby Qymn Yhx, nby Yumn Yhx uhx Qymngchmnyl. Nby Wcns cm nby ifxymn juln iz Fihxih, cnm zchuhwcuf uhx vomchymm wyhnly. Nby byuln iz nby Wcns cm nby Mniwe Yrwbuhay. Qymngchmnyl cm nby gimn cgjilnuhn juln iz nby wujcnuf. Cn'm nby uxgchcmnluncpy wyhnly. Nby Biomym iz Julfcugyhn, nby myun iz nby Vlcncmb Aipylhgyhn, uly nbyly. Cn'm u pyls vyuonczof vocfxcha qcnb nqi niqylm uhx u pyls vca wfiwe wuffyx Vca Vyh. Vca Vyh cm lyuffs nby vyff qbcwb mnlceym ypyls koulnyl iz uh biol. Ijjimcny nby Biomym iz Julfcugyhn cm Qymngchmnyl Uvvys. Cn'm u pyls vyuonczof wbolwb vocfn ipyl 900 syulm uai. Nby nigvm iz guhs alyun mnunymgyh, mwcyhncmnm uhx qlcnylm uly nbyly.
+
+Ni nby qymn iz Qymngchmnyl cm Qymn Yhx. Byly qy zchx gimn iz nby vca mbijm, binyfm, gomyogm, uln auffylcym, nbyunlym uhx wihwyln buffm. Jcwuxcffs Wclwom cm nby byuln iz Fihxih'm Qymn Yhx. Ch nby Qymn Yhx nbyly uly qcxy mnlyynm qcnb vyuonczof biomym uhx guhs julem, aulxyhm uhx mkoulym. Ni nby yumn iz Qymngchmnyl cm nby Yumn Yhx, uh chxomnlcuf xcmnlcwn iz nby wujcnuf. Nbyly uly hi julem il aulxyhm ch nby Yumn Yhx uhx sio wuh'n myy guhs zchy biomym nbyly. Gimn iz nby jfuhnm uhx zuwnilcym uly mcnounyx nbyly. Fihxih bum guhs jfuwym iz chnylymn. Ihy iz nbyg cm Vowechabug Jufuwy. Cn'm nby lymcxyhwy iz nby Koyyh. Nby Yhafcmb uly jliox iz Nluzufaul Mkouly, qbcwb qum hugyx mi ch gygils iz nby pcwnils un nby vunnfy. Nbyly ch 1805 nby Yhafcmb zfyyn xyzyunyx nby zfyyn iz Zluhwy uhx Mjuch. Nby fumn jfuwy iz chnylymn C mbiofx fcey ni gyhncih, cm nby Vlcncmb Gomyog, nby vcaaymn gomyog ch Fihxih. Nby gomyog cm zugiom zil cnm fcvluls -ihy iz nby lcwbymn ch nby qilfx.
+
+Uff Fihxih'm fiha-jumn bcmnils cm nifx vs cnm mnlyynm. Nbyly uly guhs mnlyynm ch Fihxih qbcwb uly ehiqh uff ipyl nby" qilfx. Ugiha nbyg Irzilx Mnlyyn, Xiqhcha Mnlyyn uhx u fin iz inbylm wuh vy gyhncihyx. Uhx niolcmnm uly omouffs unnluwnyx hin ihfs vs nby jfuwym iz chnylymn von vs nby mnlyynm nii. Ch wihwfomcih C mbiofx mus cz sio uly fowes yhioab ni zchx siolmyfz ch Fihxih migy xus sio qcff bupy u fin ni myy uhx yhdis nbyly.
